Whitespace and indentation changes
It's somewhat unidiomatic to use 8 space indent in Haskell code.
Typically, 4 or 2 space are common. I like four space, but two space
works great too.

When doing a hanging `do`, it's best to indent it only four spaces. That
makes the code a bit clearer to read, and less indented off the side.