Argument parsing
Truthfully, the proper way to do this is to incorporate
`optparse-generic` or `optparse-applicative` or similar to parse the
command line arguments. However, this is a little improvement.

Instead of using unsafe functions like `init`, `tail`, `last`, etc, we
pattern match on the `args` list. If we don't have at least the
secondFile and firstFile in the arguments list, then we error with a
usage message.