Safety improvements
This commit removes the unsafe `tail` and `head` functions -- note that
`head []` and `tail []` will both blow up at runtime with extremely
uninformative error messages. We should try to avoid them wherever
possible.

The replacement function folds over the list of arguments, and does a
pattern match on each one. If the pattern matches, then we grab the rest
and concatenate it with the accumulator. If the pattern doesn't match,
we just return the accumulator unchanged.