Resource safety
This isn't strictly necessary for this example, but it is in larger
programs. The pattern `openFile` requires that you *close* the file at
some point. This is easy to forget to do. It's also easy to miss some
exception behavior which leaves the files open accidentally.

Instead of:

    file <- openFile "some-file.txt" ReadMode

it is better to write:

    withFile "some-file.txt" ReadMode $ \file -> do

The function `withFile` will automatically close the file if an
exception is raised, and you don't have to remember to close the file
out -- the function will handle it for you.